// apps/console/src/__generated__/core/RiskScenariosPageQuery.graphql.ts
export type RiskScenariosPageQuery$variables = {
  riskId: string;
};
export type RiskScenariosPageQuery$data = {
  readonly node: {
    readonly __typename: "Risk";
    readonly id: string;
    readonly scenarios: {
      readonly __id: string;
      readonly edges: ReadonlyArray<{
        readonly node: {
          readonly description: string | null | undefined;
          readonly id: string;
          readonly name: string;
          readonly scope: {
            readonly riskAssessmentId: string;
          } | null | undefined;
        };
      }>;
    };
  } | {
    // This will never be '%other', but we need some
    // value in case none of the concrete values match.
    readonly __typename: "%other";
  };
};
export type RiskScenariosPageQuery = {
  response: RiskScenariosPageQuery$data;
  variables: RiskScenariosPageQuery$variables;
};